save - task 8 runs

+++ 0x04-signals/8-signal_send.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
-PID=$1
-kill -SIGQUIT $PID
+if [ "$#" -ne 1 ];
+then
+    echo "Usage: $0 <pid>"
+else
+    PID=$1
+    kill -SIGQUIT $PID
+fi
